requirements.txt
fastapi~=0.115.12
uvicorn[standard]
httpx~=0.28.1
python-dotenv~=1.1.0
pika~=1.3.2
psycopg2-binary~=2.9.10
numpy~=2.2.5
pydantic~=2.11.4
openai~=1.77.0
fastmcp>=2.3.0
pytest~=8.3.5